adds initial version of parser
Using tree sitter, this lib extract the AST of a source file. Then I
implemented a quick and dirty (using claude, with a lot of fixes,
not sure I gain a lot of time) parser.
This is "waorking" for python, I will need to do more tests though,
and make it work for other languages, and refactor, refactor, ...

Also did a dirty way to give the chunks to the indexer. The indexer
is supposed to run as a daemon, and we should send chunks using
stdin, this is not what is done here, we are just feeding the chunks
of a single file, waiting a bit and closing everything.

Finally did a quick "query" script in python to query the database
to see if we get interesting results... Will need to add more files,
and also check model that might be more suited for this type of
content.