use chroma server instead of embedded [BROKEN]
As the embedded version was creating a lot of issue and I suspect
this was due to indexing with multiple threads on the same file, the
server might provide more guard rail for those operations.

We were also having an issue with rate limitation by hugging face,
the model were loaded every time 😳

Using claude, I added two scripts, one to manage model locally and
one to manage chroma server.

Then we need to do a couple of things:
- try to start chroma server when launching `mm` (if already started
  we don't do nothing)
- maybe download the model the first time we run `mm` and then we
  check if the model is available (we need something really fast, or
maybe we assume it is there and if it fail, we download and retry)
- FIX THE TESTS! as we will need a running chroma server